Leveraging -e switch to grep for multiple exclude patterns
This change makes it easier to add to the *textmate-gf-exclude*
variable (now renamed to *textmate-gf-excludes*) by making it a list
of patterns to be excluded rather than one giant regex (helpful while
modifying the code directly or if *textmate-gf-excludes* ever moves to
defcustom).

Originally I just wanted to add the ~$ pattern to the existing
excludes but couldn't get it to work (despite the fact that my regex
was working from the command line).